+++ rector.php
@@ -1,0 +1,15 @@
+<?php
+
+use MSpirkov\Yii2\Rector\Rules\AddPropertyTagsRector;
+use MSpirkov\Yii2\Rector\Rules\RemoveRedundantPropertyTagsRector;
+use Rector\Config\RectorConfig;
+
+return RectorConfig::configure()
+    ->withPaths([
+        __DIR__ . '/src',
+        __DIR__ . '/tests',
+    ])
+    ->withRules([
+        AddPropertyTagsRector::class,
+        RemoveRedundantPropertyTagsRector::class,
+    ]);