+++ src/drivers/db/Queue.php
@@ -23,7 +23,7 @@ use yii\queue\interfaces\StatisticsProviderInterface;
 /**
  * Db Queue.
  *
- * @property-read StatisticsProvider $statisticsProvider
+ * @property-read StatisticsInterface $statisticsProvider
  *
  * @author Roman Zhuravlev <zhuravljov@gmail.com>
  */