+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Yii2 Queue Extension Change Log
 
 3.0.0 under development
 ---
+
 - Enh #542: The minimum supported PHP version is 8.3 (@s1lver)
 - Enh #503, #546: Added PHPStan for static code analysis. Error level set to 8 (@s1lver)
 - Enh #503: Added strict typing (@s1lver)
-- Enh #542: The `pda/pheanstalk` package of the Beanstalk driver has been updated to version 8.* (@s1lver)
+- Enh #542, #556: The `pda/pheanstalk` package of the Beanstalk driver has been updated to version 8.* (@s1lver)
 - Enh #503: Removed deprecated interfaces `RetryableJob`, `Job`, `Serializer` (@s1lver)
 - Enh #503: Removed deprecated classes `Signal`, `Verbose` (@s1lver)
 - Enh #503: Deprecated driver amqp has been removed (@s1lver)
@@ -15,6 +16,9 @@ Yii2 Queue Extension Change Log
 - Enh #503: The `opis/closure` package did not support PHP 8.1 and was replaced by the `laravel/serializable-closure` package (@s1lver)
 - Enh #544: Applying Yii2 coding standards (@s1lver)
 - Bug #552: Move signal handling from AMQP Interop queue component to controller (@snewer)
+- Bug #563: Fix `@property` annotations in `InfoAction` and `Queue` drivers (mspirkov)
+- Enh #564: Allow `symfony/process` `^8.0` (lukascernydis)
+- Enh #565: Add the missing `@property` tags (mspirkov)
 
 2.3.8 January 08, 2026
 ----------------------